fix/ clean up pay with card flow
